replicate_deployment/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.21.0
easyocr>=1.7.0
replicate>=0.15.0
opencv-python-headless>=4.5.0

